+++ lib/commercetools-api/src/Client/Resource/ByProjectKeyProductProjectionsByIDGet.php
@@ -30,6 +30,7 @@ use Psr\Http\Message\ResponseInterface;
 /**
 
  * @psalm-suppress PropertyNotSetInConstructor
+ * @template-implements ProjectionSelecting<ByProjectKeyProductProjectionsByIDGet>
  * @template-implements PriceSelecting<ByProjectKeyProductProjectionsByIDGet>
  * @template-implements LocaleProjecting<ByProjectKeyProductProjectionsByIDGet>
  * @template-implements StoreProjecting<ByProjectKeyProductProjectionsByIDGet>
@@ -37,7 +38,7 @@ use Psr\Http\Message\ResponseInterface;
  * @template-implements Errorable<ByProjectKeyProductProjectionsByIDGet>
  * @template-implements Deprecatable200<ByProjectKeyProductProjectionsByIDGet>
  */
-class ByProjectKeyProductProjectionsByIDGet extends ApiRequest implements PriceSelecting, LocaleProjecting, StoreProjecting, Expandable, Errorable, Deprecatable200
+class ByProjectKeyProductProjectionsByIDGet extends ApiRequest implements ProjectionSelecting, PriceSelecting, LocaleProjecting, StoreProjecting, Expandable, Errorable, Deprecatable200
 {
     /**
      * @param ?object|array|string $body