+++ lib/commercetools-api/src/Models/Product/ProductProjection.php
@@ -17,6 +17,7 @@ use Commercetools\Api\Models\State\StateReference;
 use Commercetools\Api\Models\TaxCategory\TaxCategoryReference;
 use Commercetools\Base\DateTimeImmutableCollection;
 use Commercetools\Base\JsonObject;
+use DateTimeImmutable;
 
 interface ProductProjection extends BaseResource
 {
@@ -38,9 +39,10 @@ interface ProductProjection extends BaseResource
     public const FIELD_TAX_CATEGORY = 'taxCategory';
     public const FIELD_STATE = 'state';
     public const FIELD_REVIEW_RATING_STATISTICS = 'reviewRatingStatistics';
+    public const FIELD_PRICE_MODE = 'priceMode';
 
     /**
-     * <p>The unique ID of the Product.</p>
+     * <p>Unique identifier of the <a href="ctp:api:type:Product">Product</a>.</p>
      *
 
      * @return null|string
@@ -48,7 +50,7 @@ interface ProductProjection extends BaseResource
     public function getId();
 
     /**
-     * <p>The current version of the Product.</p>
+     * <p>Current version of the <a href="ctp:api:type:Product">Product</a>.</p>
      *
 
      * @return null|int
@@ -56,7 +58,7 @@ interface ProductProjection extends BaseResource
     public function getVersion();
 
     /**
-     * <p>User-specific unique identifier of the Product.</p>
+     * <p>User-defined unique identifier of the <a href="ctp:api:type:Product">Product</a>.</p>
      *
 
      * @return null|string
@@ -64,31 +66,58 @@ interface ProductProjection extends BaseResource
     public function getKey();
 
     /**
+     * <p>Date and time (UTC) the ProductProjection was initially created.</p>
+     *
+
+     * @return null|DateTimeImmutable
+     */
+    public function getCreatedAt();
+
+    /**
+     * <p>Date and time (UTC) the ProductProjection was last updated.</p>
+     *
+
+     * @return null|DateTimeImmutable
+     */
+    public function getLastModifiedAt();
+
+    /**
+     * <p>The <a href="ctp:api:type:ProductType">ProductType</a> defining the Attributes of the <a href="ctp:api:type:Product">Product</a>.</p>
+     *
 
      * @return null|ProductTypeReference
      */
     public function getProductType();
 
     /**
+     * <p>Name of the <a href="ctp:api:type:Product">Product</a>.</p>
+     *
 
      * @return null|LocalizedString
      */
     public function getName();
 
     /**
+     * <p>Description of the <a href="ctp:api:type:Product">Product</a>.</p>
+     *
 
      * @return null|LocalizedString
      */
     public function getDescription();
 
     /**
+     * <p>User-defined identifier used in a deep-link URL for the <a href="ctp:api:type:Product">Product</a>.
+     * Must be unique across a Project, but can be the same for Products in different locales.
+     * Matches the pattern <code>[a-zA-Z0-9_-]{2,256}</code>.
+     * For <a href="/../api/predicates/query#performance-considerations">good performance</a>, indexes are provided for the first 15 <code>languages</code> set in the <a href="ctp:api:type:Project">Project</a>.</p>
+     *
 
      * @return null|LocalizedString
      */
     public function getSlug();
 
     /**
-     * <p>References to categories the product is in.</p>
+     * <p><a href="ctp:api:type:Category">Categories</a> assigned to the <a href="ctp:api:type:Product">Product</a>.</p>
      *
 
      * @return null|CategoryReferenceCollection
@@ -96,78 +125,108 @@ interface ProductProjection extends BaseResource
     public function getCategories();
 
     /**
+     * <p>Order of <a href="ctp:api:type:Product">Product</a> in <a href="ctp:api:type:Category">Categories</a>.</p>
+     *
 
      * @return null|CategoryOrderHints
      */
     public function getCategoryOrderHints();
 
     /**
+     * <p>Title of the <a href="ctp:api:type:Product">Product</a> displayed in search results.</p>
+     *
 
      * @return null|LocalizedString
      */
     public function getMetaTitle();
 
     /**
+     * <p>Description of the <a href="ctp:api:type:Product">Product</a> displayed in search results below the meta title.</p>
+     *
 
      * @return null|LocalizedString
      */
     public function getMetaDescription();
 
     /**
+     * <p>Keywords that give additional information about the <a href="ctp:api:type:Product">Product</a> to search engines.</p>
+     *
 
      * @return null|LocalizedString
      */
     public function getMetaKeywords();
 
     /**
+     * <p>Used by <a href="/../api/projects/products-suggestions">Product Suggestions</a>, but is also considered for a <a href="ctp:api:type:FullTextSearch">full text search</a>.</p>
+     *
 
      * @return null|SearchKeywords
      */
     public function getSearchKeywords();
 
     /**
+     * <p><code>true</code> if the staged data is different from the current data.</p>
+     *
 
      * @return null|bool
      */
     public function getHasStagedChanges();
 
     /**
+     * <p><code>true</code> if the <a href="ctp:api:type:Product">Product</a> is <a href="ctp:api:type:CurrentStaged">published</a>.</p>
+     *
 
      * @return null|bool
      */
     public function getPublished();
 
     /**
+     * <p>The Master Variant of the <a href="ctp:api:type:Product">Product</a>.</p>
+     *
 
      * @return null|ProductVariant
      */
     public function getMasterVariant();
 
     /**
+     * <p>Additional Product Variants.</p>
+     *
 
      * @return null|ProductVariantCollection
      */
     public function getVariants();
 
     /**
+     * <p>The <a href="ctp:api:type:TaxCategory">TaxCategory</a> of the <a href="ctp:api:type:Product">Product</a>.</p>
+     *
 
      * @return null|TaxCategoryReference
      */
     public function getTaxCategory();
 
     /**
+     * <p><a href="ctp:api:type:State">State</a> of the <a href="ctp:api:type:Product">Product</a>.</p>
+     *
 
      * @return null|StateReference
      */
     public function getState();
 
     /**
-     * <p>Statistics about the review ratings taken into account for this product.</p>
+     * <p>Review statistics of the <a href="ctp:api:type:Product">Product</a>.</p>
      *
 
      * @return null|ReviewRatingStatistics
      */
     public function getReviewRatingStatistics();
+
+    /**
+     * <p>Indicates whether the Prices of the Product Projection are <a href="ctp:api:type:Price">embedded</a> or <a href="ctp:api:type:StandalonePrice">standalone</a>. <a href="#prices">Projecting Prices</a> only works with <code>Embedded</code>, there is currently no support for <code>Standalone</code>.</p>
+     *
+
+     * @return null|string
+     */
+    public function getPriceMode();
 
     /**
      * @param ?string $id
@@ -183,6 +242,16 @@ interface ProductProjection extends BaseResource
      * @param ?string $key
      */
     public function setKey(?string $key): void;
+
+    /**
+     * @param ?DateTimeImmutable $createdAt
+     */
+    public function setCreatedAt(?DateTimeImmutable $createdAt): void;
+
+    /**
+     * @param ?DateTimeImmutable $lastModifiedAt
+     */
+    public function setLastModifiedAt(?DateTimeImmutable $lastModifiedAt): void;
 
     /**
      * @param ?ProductTypeReference $productType
@@ -268,4 +337,9 @@ interface ProductProjection extends BaseResource
      * @param ?ReviewRatingStatistics $reviewRatingStatistics
      */
     public function setReviewRatingStatistics(?ReviewRatingStatistics $reviewRatingStatistics): void;
+
+    /**
+     * @param ?string $priceMode
+     */
+    public function setPriceMode(?string $priceMode): void;
 }