+++ lib/commercetools-api/src/Models/Product/ProductProjectionBuilder.php
@@ -167,7 +167,13 @@ final class ProductProjectionBuilder implements Builder
     private $reviewRatingStatistics;
 
     /**
-     * <p>The unique ID of the Product.</p>
+
+     * @var ?string
+     */
+    private $priceMode;
+
+    /**
+     * <p>Unique identifier of the <a href="ctp:api:type:Product">Product</a>.</p>
      *
 
      * @return null|string
@@ -178,7 +184,7 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
-     * <p>The current version of the Product.</p>
+     * <p>Current version of the <a href="ctp:api:type:Product">Product</a>.</p>
      *
 
      * @return null|int
@@ -189,6 +195,8 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
+     * <p>Date and time (UTC) the ProductProjection was initially created.</p>
+     *
 
      * @return null|DateTimeImmutable
      */
@@ -198,6 +206,8 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
+     * <p>Date and time (UTC) the ProductProjection was last updated.</p>
+     *
 
      * @return null|DateTimeImmutable
      */
@@ -207,7 +217,7 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
-     * <p>User-specific unique identifier of the Product.</p>
+     * <p>User-defined unique identifier of the <a href="ctp:api:type:Product">Product</a>.</p>
      *
 
      * @return null|string
@@ -218,6 +228,8 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
+     * <p>The <a href="ctp:api:type:ProductType">ProductType</a> defining the Attributes of the <a href="ctp:api:type:Product">Product</a>.</p>
+     *
 
      * @return null|ProductTypeReference
      */
@@ -227,6 +239,8 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
+     * <p>Name of the <a href="ctp:api:type:Product">Product</a>.</p>
+     *
 
      * @return null|LocalizedString
      */
@@ -236,6 +250,8 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
+     * <p>Description of the <a href="ctp:api:type:Product">Product</a>.</p>
+     *
 
      * @return null|LocalizedString
      */
@@ -245,6 +261,11 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
+     * <p>User-defined identifier used in a deep-link URL for the <a href="ctp:api:type:Product">Product</a>.
+     * Must be unique across a Project, but can be the same for Products in different locales.
+     * Matches the pattern <code>[a-zA-Z0-9_-]{2,256}</code>.
+     * For <a href="/../api/predicates/query#performance-considerations">good performance</a>, indexes are provided for the first 15 <code>languages</code> set in the <a href="ctp:api:type:Project">Project</a>.</p>
+     *
 
      * @return null|LocalizedString
      */
@@ -254,7 +275,7 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
-     * <p>References to categories the product is in.</p>
+     * <p><a href="ctp:api:type:Category">Categories</a> assigned to the <a href="ctp:api:type:Product">Product</a>.</p>
      *
 
      * @return null|CategoryReferenceCollection
@@ -265,6 +286,8 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
+     * <p>Order of <a href="ctp:api:type:Product">Product</a> in <a href="ctp:api:type:Category">Categories</a>.</p>
+     *
 
      * @return null|CategoryOrderHints
      */
@@ -274,6 +297,8 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
+     * <p>Title of the <a href="ctp:api:type:Product">Product</a> displayed in search results.</p>
+     *
 
      * @return null|LocalizedString
      */
@@ -283,6 +308,8 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
+     * <p>Description of the <a href="ctp:api:type:Product">Product</a> displayed in search results below the meta title.</p>
+     *
 
      * @return null|LocalizedString
      */
@@ -292,6 +319,8 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
+     * <p>Keywords that give additional information about the <a href="ctp:api:type:Product">Product</a> to search engines.</p>
+     *
 
      * @return null|LocalizedString
      */
@@ -301,6 +330,8 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
+     * <p>Used by <a href="/../api/projects/products-suggestions">Product Suggestions</a>, but is also considered for a <a href="ctp:api:type:FullTextSearch">full text search</a>.</p>
+     *
 
      * @return null|SearchKeywords
      */
@@ -310,6 +341,8 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
+     * <p><code>true</code> if the staged data is different from the current data.</p>
+     *
 
      * @return null|bool
      */
@@ -319,6 +352,8 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
+     * <p><code>true</code> if the <a href="ctp:api:type:Product">Product</a> is <a href="ctp:api:type:CurrentStaged">published</a>.</p>
+     *
 
      * @return null|bool
      */
@@ -328,6 +363,8 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
+     * <p>The Master Variant of the <a href="ctp:api:type:Product">Product</a>.</p>
+     *
 
      * @return null|ProductVariant
      */
@@ -337,6 +374,8 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
+     * <p>Additional Product Variants.</p>
+     *
 
      * @return null|ProductVariantCollection
      */
@@ -346,6 +385,8 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
+     * <p>The <a href="ctp:api:type:TaxCategory">TaxCategory</a> of the <a href="ctp:api:type:Product">Product</a>.</p>
+     *
 
      * @return null|TaxCategoryReference
      */
@@ -355,6 +396,8 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
+     * <p><a href="ctp:api:type:State">State</a> of the <a href="ctp:api:type:Product">Product</a>.</p>
+     *
 
      * @return null|StateReference
      */
@@ -364,7 +407,7 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
-     * <p>Statistics about the review ratings taken into account for this product.</p>
+     * <p>Review statistics of the <a href="ctp:api:type:Product">Product</a>.</p>
      *
 
      * @return null|ReviewRatingStatistics
@@ -372,6 +415,17 @@ final class ProductProjectionBuilder implements Builder
     public function getReviewRatingStatistics()
     {
         return $this->reviewRatingStatistics instanceof ReviewRatingStatisticsBuilder ? $this->reviewRatingStatistics->build() : $this->reviewRatingStatistics;
+    }
+
+    /**
+     * <p>Indicates whether the Prices of the Product Projection are <a href="ctp:api:type:Price">embedded</a> or <a href="ctp:api:type:StandalonePrice">standalone</a>. <a href="#prices">Projecting Prices</a> only works with <code>Embedded</code>, there is currently no support for <code>Standalone</code>.</p>
+     *
+
+     * @return null|string
+     */
+    public function getPriceMode()
+    {
+        return $this->priceMode;
     }
 
     /**
@@ -617,6 +671,17 @@ final class ProductProjectionBuilder implements Builder
     }
 
     /**
+     * @param ?string $priceMode
+     * @return $this
+     */
+    public function withPriceMode(?string $priceMode)
+    {
+        $this->priceMode = $priceMode;
+
+        return $this;
+    }
+
+    /**
      * @deprecated use withProductType() instead
      * @return $this
      */
@@ -783,7 +848,8 @@ final class ProductProjectionBuilder implements Builder
             $this->variants,
             $this->taxCategory instanceof TaxCategoryReferenceBuilder ? $this->taxCategory->build() : $this->taxCategory,
             $this->state instanceof StateReferenceBuilder ? $this->state->build() : $this->state,
-            $this->reviewRatingStatistics instanceof ReviewRatingStatisticsBuilder ? $this->reviewRatingStatistics->build() : $this->reviewRatingStatistics
+            $this->reviewRatingStatistics instanceof ReviewRatingStatisticsBuilder ? $this->reviewRatingStatistics->build() : $this->reviewRatingStatistics,
+            $this->priceMode
         );
     }